Makes repo Android-build-ready but ChangeScene script unfinished
The tab system is a work in-progress as it is still not fully functional. Note that only the first scene has been added to the build, so touching the 2nd and 3rd tabs do nothing. The first tab simply resets the FirstLaw scene, but we aim to change scenes not fully reset scenes in the final version. Note that the names of the tabs are unclickable and only the colored area of the tabs are clickable as of now.